chore(examples) use react in mapbox website example
