+++ examples/website/mapbox/index.html
@@ -22,7 +22,7 @@
     <div id="app"></div>
   </body>
   <script type="module">
-    import {loadAndRender} from './app.js';
-    loadAndRender(document.getElementById('app'));
+    import {renderToDOM} from './app.jsx';
+    renderToDOM(document.getElementById('app'));
   </script>
 </html>
