+++ examples/website/mapbox/package.json
@@ -12,7 +12,10 @@
     "@loaders.gl/csv": "^4.1.0",
     "d3-scale": "^2.0.0",
     "deck.gl": "^9.0.0-beta",
-    "mapbox-gl": "^3.0.0"
+    "mapbox-gl": "^3.0.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-map-gl": "^7.0.0"
   },
   "devDependencies": {
     "typescript": "^4.6.0",
